+++ simple_detrender.cpp
@@ -12,17 +12,17 @@ namespace rf_pipelines {
 
 
 struct simple_detrender : public wi_transform {
-    simple_detrender(int nt_chunk_);
+    simple_detrender(ssize_t nt_chunk_);
 
     virtual void set_stream(const wi_stream &stream);
     virtual void start_substream(int isubstream, double t0) { }
     virtual void end_substream() { cerr << "simple_detrender::end_substream() called!\n"; }
 
-    virtual void process_chunk(double t0, double t1, float *intensity, float *weights, int stride, float *pp_intensity, float *pp_weights, int pp_stride);
+    virtual void process_chunk(double t0, double t1, float *intensity, float *weights, ssize_t stride, float *pp_intensity, float *pp_weights, ssize_t pp_stride);
 };
 
 
-simple_detrender::simple_detrender(int nt_chunk_)
+simple_detrender::simple_detrender(ssize_t nt_chunk_)
 {
     // Note that 'nt_prepad' and nt_postpad are automatically initialized to zero.
     // We still need to initialize 'nfreq'.  This will be done in set_stream().
@@ -39,13 +39,13 @@ void simple_detrender::set_stream(const wi_stream &stream)
 }
 
 
-void simple_detrender::process_chunk(double t0, double t1, float *intensity, float *weights, int stride, float *pp_intensity, float *pp_weights, int pp_stride)
+void simple_detrender::process_chunk(double t0, double t1, float *intensity, float *weights, ssize_t stride, float *pp_intensity, float *pp_weights, ssize_t pp_stride)
 {
-    for (int ifreq = 0; ifreq < this->nfreq; ifreq++) {
+    for (ssize_t ifreq = 0; ifreq < this->nfreq; ifreq++) {
 	float num = 0.0;  // sum of weighted intensities
 	float den = 0.0;  // sum of weights
 
-	for (int it = 0; it < this->nt_chunk; it++) {
+	for (ssize_t it = 0; it < this->nt_chunk; it++) {
 	    num += weights[ifreq*stride + it] * intensity[ifreq*stride + it];
 	    den += weights[ifreq*stride + it];
 	}
@@ -55,13 +55,13 @@ void simple_detrender::process_chunk(double t0, double t1, float *intensity, flo
 	    continue;
 
 	// Note that (num/den) is the weighted mean of all the intensity samples.
-	for (int it = 0; it < this->nt_chunk; it++)
+	for (ssize_t it = 0; it < this->nt_chunk; it++)
 	    intensity[ifreq*stride + it] -= (num/den);
     }
 }
 
 // Externally visible factory function
-shared_ptr<wi_transform> make_simple_detrender(int nt_chunk)
+shared_ptr<wi_transform> make_simple_detrender(ssize_t nt_chunk)
 {
     return make_shared<simple_detrender> (nt_chunk);
 }
